packages/mediacenter/OutFox: add litboard support
First of, the Sextet FIFO is handeld by systemd so that it the
accompanying service ensures the FIFO as a reader (preventing failure to
open the file) and that it is consumed, so that OutFox doesn't block
should the reader be slow.

The service then uses socat to forward the data to UDP datagrams on
localhost:5001. Meaning that that data is lost if no liboard is plugged,
but the FIFO is always consumed.

On the other side, a udev rule on idVendor=1a86 idProduct=7523 will
start a consuming service that will fetch datagrams from :5001 and
forward them to the litboard, again using socat.

This effectively brings plug and play support of the litboard, as OutFox
is running.

Signed-off-by: Steeve Morin <steeve.morin@gmail.com>